requirements.txt
streamlit
easyocr
Pillow
pandas
gspread
oauth2client
